[spotify] Convert to sync interface + add seek support
